pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel", "setuptools-scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "node-fdm"
description = "Neural ODE model for aircraft flight dynamics"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENCE.md" }
authors = [
  { name = "Gabriel JARRY", email = "gabriel.jarry@eurocontrol.int" },
  { name = "Xavier OLIVE", email = "xavier.olive@onera.fr" },
  { name = "Ramon DALMAU", email = "ramon.dalmau-codina@eurocontrol.int" },
  { name = "Philippe VERY", email = "philippe.very@eurocontrol.int" }
]

dynamic = ["version"]

dependencies = [
  "torch>=2.6.0",
  "torchdiffeq>=0.2.5",
  "numpy>=1.26",
  "pandas>=2.2.3",
  "scipy>=1.13.1",
  "pyarrow>=21.0",
  "joblib>=1.4.2",
  "matplotlib>=3.10.1",
  "tqdm>=4.67.1",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
all = [
  "traffic>=2.10.2",
  "fastmeteo>=1.1.0",
  "click>=8.1.8",
  "altair>=5.5",
  "cairosvg>=2.7",
  "vl-convert-python>=1.8.0",
  "ipykernel>=6.29.5",
]

bada = [
  "pyBADA>=0.1.7",
]

test = [
  "pytest>=7.4",
]

docs = [
  "mkdocs-material>=9",
  "mkdocstrings[python]>=0.27",
]

[project.urls]
Homepage = "https://github.com/eurocontrol"

[tool.setuptools_scm]
write_to = "node_fdm/_version.py"
fallback_version = "0.0.0-dev"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = [
  "node_fdm*",
  "code*",
  "preprocessing*",
  "pybada_predictor*",
  "utils*",
]
